agent-hiring-mvp/agenthub-sdk/templates/research_team_expertise_agent/requirements.txt
# Core dependencies
requests>=2.25.1,<3.0.0
python-dotenv>=0.19.0,<2.0.0

# Academic data extraction
arxiv>=1.4.8,<2.0.0
scholarly>=1.7.11,<2.0.0
semanticscholar>=0.10.0,<1.0.0

# Web scraping and search
beautifulsoup4>=4.9.3,<5.0.0
lxml>=4.6.3,<5.0.0
duckduckgo-search>=4.1.1,<5.0.0

# Data processing
pandas>=1.3.0,<3.0.0
numpy>=1.21.0,<2.0.0
networkx>=2.6.0,<4.0.0

# AI/ML (if needed)
openai>=1.0.0,<2.0.0

# LangChain dependencies
langchain>=0.1.0,<1.0.0
langchain-openai>=0.1.0,<1.0.0
langchain-community>=0.1.0,<1.0.0

# Vector store
faiss-cpu>=1.7.0,<2.0.0

# Azure OpenAI support via LiteLLM
litellm>=1.0.0,<2.0.0

# Note: serper-api is optional and may cause conflicts
# Install manually if needed: pip install serper-api
